src/use-cases/fetch-user-check-in-history/fetch-user-check-in-history.ts
import { CheckIn } from '@prisma/client'
import { ResourceNotFoundError } from '../errors/resource-not-found'
import { CheckInsRepository } from '@/repositories/check-in-repository'

interface FetchUserCheckInHistoryUseCaseRequest {
  userId: string
  page: number
}

interface FetchUserCheckInHistoryUseCaseResponse {
  checkIns: CheckIn[]
}

export class FetchUserCheckInHistoryUseCase {
  constructor(private checkInsRepository: CheckInsRepository) {}

  async execute({
    userId,
    page,
  }: FetchUserCheckInHistoryUseCaseRequest): Promise<FetchUserCheckInHistoryUseCaseResponse> {
    const checkIns = await this.checkInsRepository.findManyByUserId(
      userId,
      page,
    )

    if (!checkIns) {
      throw new ResourceNotFoundError('Check-ins')
    }

    return {
      checkIns,
    }
  }
}
